Refactor: 파일명 수정 + modal 컴포넌트 위치 변경#50
